``` feat(alipay): 集成支付宝支付功能并优化配置
- 添加支付宝相关配置类AlipayConfig，实现InitializingBean接口用于初始化日志
- 集成支付宝SDK，添加支付页面生成、异步通知处理等核心功能
- 实现支付结果验证、订单状态查询等功能
- 配置支付宝沙箱环境参数，包括应用ID、密钥、网关等信息
- 添加支付成功后的业务逻辑处理，支持罚金和押金支付场景

refactor: 统一项目模块名称从library-ssm到library

- 修改pom.xml中的artifactId和name为library
- 更新web.xml中的display-name- 修改IDEA编译器配置中的模块名称
- 更新日志文件路径从library-ssm.log到library.log
- 调整Maven打包相关配置chore: 升级Java版本至JDK21并清理配置文件

- 将Maven编译目标版本从JDK17升级至JDK 21
- 移除IDEA中不必要的数据源配置文件
- 清理项目根目录下的配置文件
-修复CSS样式中的百分号转义问题

feat(borrow): 增加借阅状态显示功能

- 在图书详情页添加当前用户借阅状态显示- 实现currentBorrowing方法查询用户的当前借阅记录
- 在页面控制器中注入借阅状态数据到模型

fix(common): 修复Redis缓存服务中的参数类型错误-修正CacheService中所有Redis操作的过期时间参数类型
- 将int类型的过期时间统一转换为long类型以避免精度丢失refactor(common): 统一前端错误提示方式

- 将JavaScript中的alert弹窗替换为Toast组件
- 改进错误提示的用户体验```

+++ src/main/java/com/library/service/CacheService.java
@@ -53,7 +53,7 @@ public class CacheService {
     /** 写入缓存（JSON 序列化 + 过期时间，单位秒） */
     public void set(String key, Object value, int ttlSeconds) {
         try (Jedis jedis = jedisPool.getResource()) {
-            jedis.setex(key, ttlSeconds, OBJECT_MAPPER.writeValueAsString(value));
+            jedis.setex(key, (long) ttlSeconds, OBJECT_MAPPER.writeValueAsString(value));
         } catch (Exception e) {
             log.warn("Redis set 降级, key={}, 原因: {}", key, e.getMessage(), e);
         }
@@ -81,7 +81,7 @@ public class CacheService {
     /** 设置过期时间（秒） */
     public void expire(String key, int ttlSeconds) {
         try (Jedis jedis = jedisPool.getResource()) {
-            jedis.expire(key, ttlSeconds);
+            jedis.expire(key, (long) ttlSeconds);
         } catch (Exception e) {
             log.warn("Redis expire 降级, key={}, 原因: {}", key, e.getMessage());
         }
@@ -90,7 +90,7 @@ public class CacheService {
     /** 带过期时间的 setnx（同 tryLock 语义，但返回 boolean 更通用） */
     public boolean setnx(String key, String value, int ttlSeconds) {
         try (Jedis jedis = jedisPool.getResource()) {
-            String result = jedis.set(key, value, SetParams.setParams().nx().ex(ttlSeconds));
+            String result = jedis.set(key, value, SetParams.setParams().nx().ex((long) ttlSeconds));
             return "OK".equals(result);
         } catch (Exception e) {
             log.warn("Redis setnx 降级, key={}, 原因: {}", key, e.getMessage());
@@ -153,7 +153,7 @@ public class CacheService {
             jedis.del(key);
             if (scoreMap != null && !scoreMap.isEmpty()) {
                 jedis.zadd(key, scoreMap);
-                jedis.expire(key, ttlSeconds);
+                jedis.expire(key, (long) ttlSeconds);
             }
         } catch (Exception e) {
             log.warn("Redis zadd 降级, key={}, 原因: {}", key, e.getMessage());
@@ -170,7 +170,7 @@ public class CacheService {
      */
     public boolean tryLock(String key, String token, long expireSeconds) {
         try (Jedis jedis = jedisPool.getResource()) {
-            String result = jedis.set(key, token, SetParams.setParams().nx().ex((int) expireSeconds));
+            String result = jedis.set(key, token, SetParams.setParams().nx().ex(expireSeconds));
             return "OK".equals(result);
         } catch (Exception e) {
             log.warn("Redis tryLock 降级, key={}, 原因: {}", key, e.getMessage());
